Dedupe `getChangesFileRowItems`
